+++ src/class/Middleware.ts
@@ -10,7 +10,10 @@ import {
  * and 'responseBuilder', which determines the format of the responses.
  * The 'handle' method must be implemented, even if it is not necessary in a controller.
  */
-export default abstract class Middleware<G extends GenericTypes> {
+export default abstract class Middleware<
+  G extends GenericTypes,
+  ApiResponse extends Record<string, any> = Record<string, any>
+> {
   static moduleName?: string;
 
   context: Context | null;
@@ -35,19 +38,19 @@ export default abstract class Middleware<G extends GenericTypes> {
 
   abstract handle(params: RequestHandlerParams<G>): any;
 
-  static handler(c: typeof Middleware, name: string) {
-    if (!c.moduleName) throw new Error(`Module name not defined in ${c.name}`);
-    return `${c.moduleName}.${name}`;
+  static handler<T = string>(methodName: T, moduleName?: string): string {
+    moduleName = moduleName ?? this.moduleName ?? this.name;
+    return `${moduleName}.${methodName}`;
   }
 
-  
-  
   /**
    * Builds a response from a string, number, or object.
    * @param {string | number | T} res - The response to build.
    * @returns {string | number | T} - The built response.
    */
-  responseBuilder<T>(res: string | number | T): string | number | T {
+  responseBuilder(
+    res: string | number | ApiResponse
+  ): string | number | ApiResponse {
     return res;
   }
 }